% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_mrt_mosaic.R
\name{batch_mrt_mosaic}
\alias{batch_mrt_mosaic}
\title{Mosiac a series of MODSI satellite images using MRT.}
\usage{
batch_mrt_mosaic(raw_hdf_path, mosaic_hdf_path, bands_subset = "",
  delete = FALSE, .parallel = TRUE, nCores = 4)
}
\arguments{
\item{raw_hdf_path}{The path of the downloaded hdf files.}

\item{mosaic_hdf_path}{The path to store the mosaiced hdf files.}

\item{bands_subset}{A character vector indicating the spectral subset of the input hdf file.
Pass to \code{\link{mosaic_modis_hdf}} function.}

\item{delete}{Logical. If TRUE the original hdf files will be deleted after mosaicing. Default is FALSE.
Pass to \code{\link{mosaic_modis_hdf}} function.}

\item{.parallel}{Logical. Should mosaic be executed in parallel mode?}

\item{nCores}{Numeric. The number of clusters to use when \code{.parallel = T}.}
}
\description{
Runs \code{\link{mosaic_modis_hdf}} on a batch of files.
It internally use MODIS Reproject Tool (MRT) software to mosaic the downloaded images.
}
\note{
To have the functionality for Mosaic and reprojecting of the images,
you need to first install MRT software (\url{https://lpdaac.usgs.gov/tools/modis_reprojection_tool})
on your machine, and introduce its' path through the
mrt_path argument.
}
\seealso{
\code{\link{mosaic_modis_hdf}} \code{\link{reproject_modis_hdf}} \code{\link{get_hdf_dates}}
}
