% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_mrt_reporjection.R
\name{batch_mrt_reporjection}
\alias{batch_mrt_reporjection}
\title{Reproject a series of MODSI satellite images using MRT.}
\usage{
batch_mrt_reporjection(mosaic_hdf_path, mosaic_tif_path, .parallel = T,
  nCores = 4, bands_subset = "", proj_type, rename = T, ...)
}
\arguments{
\item{mosaic_hdf_path}{The path to store the mosaiced hdf files.}

\item{mosaic_tif_path}{The path to stroe the mosaiced tif files.}

\item{.parallel}{Logical. Should reprojection be executed in parallel mode?}

\item{nCores}{Numeric. The number of clusters to use when \code{.parallel = T}.}

\item{bands_subset}{A character vector indicating the spectral subset of the input hdf file.
Pass to \code{\link{reproject_modis_hdf}} function.}

\item{proj_type}{The projection type used. Default is "UTM". Pass to \code{\link{reproject_modis_hdf}} function.}

\item{rename}{Logical. If TRUE the file will be renamed using the date of the HDF file.}

\item{...}{Other arguments pass to \code{\link{reproject_modis_hdf}} function.}
}
\description{
Runs \code{\link{reproject_modis_hdf}} on a batch of files.
It internally use MODIS Reproject Tool (MRT) software to mosaic the downloaded images.
}
\note{
To have the functionality for Mosaic and reprojecting of the images,
you need to first install MRT software (\url{https://lpdaac.usgs.gov/tools/modis_reprojection_tool})
on your machine, and introduce its' path through the
mrt_path argument.
}
\seealso{
\code{\link{mosaic_modis_hdf}} \code{\link{reproject_modis_hdf}} \code{\link{get_hdf_dates}}
}
