% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_raster_calc.R
\name{batch_raster_calc}
\alias{batch_raster_calc}
\title{Calculate batch of satellite images using raster.}
\usage{
batch_raster_calc(infiles, outdir, outsuffix = ".tif", pattern = ".tif$",
  .parallel = T, nCores = 4, .fun, recursive = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{infiles}{Character. A directory or a character vector of files (including their path).
If a directory, all files matching the pattern will be converted.}

\item{outdir}{Character. Output directory to save the output files.}

\item{outsuffix}{Character. The suffix to append to the input filename (minus its extension)
to generate the output filename(s).}

\item{pattern}{Character. If infiles is a directory, this is used to limit the file it is searching for.}

\item{.parallel}{Logical. Should mask be executed in parallel mode?}

\item{nCores}{Numeric. The number of clusters to use when \code{.parallel = T}.}

\item{.fun}{Function to apply to each raster.}

\item{recursive}{Logical. If infiles is a directory, should files be searched for recursively?}

\item{verbose}{Logical. Enable verbose execution? Default is FALSE.}

\item{...}{Parameters to pass to \code{\link[raster]{writeRaster}}.}
}
\value{
Either a list of NULLs or
a list of RasterBricks depending on whether output_Raster is set to TRUE.
}
\description{
Calculate batch of satellite images using raster.
}
\details{
This function is designed to run customized functions over rasters in batch mode.
Files are passed to the function either directly as a character vector of filenames,
or by passing it a directory and (typically) a search pattern (e.g. pattern=".tif").

If a parallel engine is started and registered with foreach,
this program will run in parallel.
\code{writeRaster} will execute based on parameters passed to it,
and the output file will be named based on the input file (stripped of its extension),
with the outsuffix appended to it.
}
\seealso{
\code{\link[raster]{writeRaster}}
}
