% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdf_dates.R
\name{get_hdf_dates}
\alias{get_hdf_dates}
\title{Get the dates of the MODIS hdf files.}
\usage{
get_hdf_dates(hdf_path, pattern = "\\\\d{4}M\\\\d{2}")
}
\arguments{
\item{hdf_path}{The path of the downloaded hdf files.}

\item{pattern}{Pattern to look for. Pass to \code{\link[stringr]{str_extract}}.}
}
\value{
A character vector.
}
\description{
Get the dates of the MODIS hdf files.
}
