% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_taylor.R
\name{gg_taylor}
\alias{gg_taylor}
\title{Plot Taylor Diagram with ggplot2}
\usage{
gg_taylor(mods, obs, normalize = T, ...)
}
\arguments{
\item{mods}{List. A named list which contain the modelled values. Model names are
extracted using the list names. Observations for each model should be the same in length.}

\item{obs}{Vector. A numeric vector which contain the observed values and the length of
which should be the same with that of the models.}

\item{normalize}{Logical. If TRUE, the standard deviation of the observations will be set
to 1 and standard deviation of the models will be normalized using the actural standard
deviation of the observations. Default is TRUE}

\item{...}{Extra arguments. Currently not implemented.}
}
\value{
A ggplot2 object. If not assigned, it will plot in the Plots panel.
}
\description{
Plot Taylor Diagram with ggplot2
}
\examples{
\dontrun{
set.seed(1234)
obs <- rnorm(100)
mods <- list(mod1 = obs +rnorm(100)*0.1,
             mod2 = obs +rnorm(100)*0.1,
             mod3 = obs+rnorm(100)/2,
             mod4 = obs+rnorm(100)/2)
gg_taylor(mods = mods, obs = obs, normalize = T)
}
}
