% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODISR_AQUA.R
\name{modisr_aqua_product_short_name}
\alias{modisr_aqua_product_short_name}
\title{Map MODIS AQUA Product Names to Short Codes}
\usage{
modisr_aqua_product_short_name(product)
}
\arguments{
\item{product}{A character string specifying the product name.
Supported product names include "MODIS AQUA L2 SST", "MODIS AQUA L3 Binned SST",
"MODIS AQUA L3 Binned CHL", and "MODIS AQUA L3 Mapped CHL".}
}
\value{
A character string of the corresponding short code for the specified product.
}
\description{
Maps user-friendly MODIS AQUA product names to their respective
short codes used in API queries. This function supports mapping for a
predefined list of products including SST (Sea Surface Temperature) and
CHL (Chlorophyll) data products.
}
\examples{
modisr_aqua_product_short_name("MODIS AQUA L2 SST")
}
