% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_modis_hdf.R
\name{mosaic_modis_hdf}
\alias{mosaic_modis_hdf}
\title{Mosaic MODIS hdf files.}
\usage{
mosaic_modis_hdf(hdf_fpnames, dst_fpname, mrt_path, bands_subset,
  delete = FALSE)
}
\arguments{
\item{hdf_fpnames}{The names of the hdf files to be mosaiced.}

\item{dst_fpname}{The full name of the output hdf file. e.g. "C:/out.hdf".}

\item{mrt_path}{The path of the MRT executable file. If missing, default is "C:/MRT/bin".}

\item{bands_subset}{A character vector indicating the spectral subset of the input hdf file.}

\item{delete}{Logical. If TRUE the original hdf files will be deleted after mosaicing.
Default is FALSE.}
}
\description{
Modified from the original version in the 'rts' pacakge by Babak Naimi.
It internally use MODIS Reproject Tool (MRT) software to mosaic the downloaded images.
}
\note{
To have the functionality for Mosaic and reprojecting of the images,
you need to first install MRT software (\url{https://lpdaac.usgs.gov/tools/modis_reprojection_tool})
on your machine, and introduce its' path through the
mrt_path argument.
}
