% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject_modis_hdf.R
\name{reproject_modis_hdf}
\alias{reproject_modis_hdf}
\title{Reproject MODIS hdf file and save it in desired format.}
\usage{
reproject_modis_hdf(hdf_fpname, dst_fpname, mrt_path, UL = "", LR = "",
  resample_type = "NEAREST_NEIGHBOR", proj_type = "UTM",
  bands_subset = "", proj_params = "0 0 0 0 0 0 0 0 0 0 0 0 0 0 0",
  datum = "WGS84", utm_zone = 50, pixel_size = 1000)
}
\arguments{
\item{hdf_fpname}{The name of the hdf file.}

\item{dst_fpname}{The full name of the output file. e.g. "C:/out.tif".}

\item{mrt_path}{The path of the MRT executable file. If missing, default is "C:/MRT/bin".}

\item{UL}{The upper-left corner of the processing extent.}

\item{LR}{The lower-right corner of the processing extent.}

\item{resample_type}{The resample method used. Default is "NEAREST_NEIGHBOR".}

\item{proj_type}{The projection type used. Default is "UTM".}

\item{bands_subset}{A character vector indicating the spectral subset of the input hdf file.}

\item{proj_params}{A character vector indicating the project parameters.}

\item{datum}{Character of datum to use. Default is "WGS84".}

\item{utm_zone}{Number of the UTM zones when projection. Only used when proj_type = "UTM".}

\item{pixel_size}{Number of the output pixel size. Only used when proj_type != "GEO".}
}
\description{
Modified from the original version in the 'rts' pacakge by Babak Naimi.
It internally use MODIS Reproject Tool (MRT) software to reproject the mosaiced image
to a specified coordinate system. As the format of the source images in LP DAAC is HDF,
this tool can also convert them into other formats (i.e. Geotif, hdr).
}
\note{
To have the functionality for Mosaic and reprojecting of the images,
you need to first install MRT software (\url{https://lpdaac.usgs.gov/tools/modis_reprojection_tool})
on your machine, and introduce its' path through the
mrt_path argument.
}
