% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_foreach_pb.R
\name{set_foreach_pb}
\alias{set_foreach_pb}
\title{Create progress bar for \code{foreach} function.}
\usage{
set_foreach_pb(max, style = 3)
}
\arguments{
\item{max}{Numeric value (finite) for the extreme of the progress bar.}

\item{style}{The style of the bar. See Details.}
}
\value{
A list which could pass to \code{\link[foreach]{foreach}}
through \code{.options.snow} argument.
}
\description{
Create progress bar for \code{foreach} function.
}
\details{
\code{style = 1} and \code{style = 2} just shows a line of char.
They differ in that \code{style = 2} redraws the line each time,
which is useful if other code might be writing to the R console.
\code{style = 3} marks the end of the range by | and gives a percentage
to the right of the bar.
}
\seealso{
\code{\link[utils]{txtProgressBar}} \code{\link[utils]{setTxtProgressBar}}
}
