% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_taylor.R
\name{theme_taylor}
\alias{theme_taylor}
\title{Themes for taylor diagram.}
\usage{
theme_taylor(base_size = 11, base_family = "serif")
}
\arguments{
\item{base_size}{Numeric. Parameter pass to \code{\link[ggplot2]{theme_bw}}.}

\item{base_family}{Character. Parameter pass to \code{\link[ggplot2]{theme_bw}}.}
}
\description{
Themes for taylor diagram.
}
