% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODISR_AQUA.R
\name{modisr_aqua_list_files}
\alias{modisr_aqua_list_files}
\title{https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html}
\usage{
modisr_aqua_list_files(
  product = "MODIS AQUA L2 SST",
  max_results = 20,
  temporal = NULL,
  bounding_box = list(n_lat = 90, s_lat = -90, w_lon = -180, e_lon = 180),
  polygon = NULL,
  time_resolution = NULL
)
}
\description{
https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html
}
